Separate out telephony and usage stats modules
