--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "commons-android"]
-	path = commons-android
-	url = https://github.com/RADAR-CNS/RADAR-Commons-Android.git
