+++ radar-android-phone-telephony/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.radarcns.phone.telephony">
+    <uses-feature android:name="android.hardware.telephony" android:required="false"/>
+
+    <uses-permission android:name="android.permission.READ_CALL_LOG"/>
+    <uses-permission android:name="android.permission.READ_SMS"/>
+
+    <application android:allowBackup="true">
+        <service android:name=".PhoneLogService"/>
+    </application>
+</manifest>
