+++ radar-android-phone-telephony/src/main/java/org/radarcns/phone/telephony/RecordProcessor.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2018 The Hyve
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.radarcns.phone.telephony;
+
+import android.database.Cursor;
+
+/**
+ * Functional interface for processing a record.
+ */
+interface RecordProcessor {
+    /**
+     * Process a single record/
+     * @param cursor current record to process
+     * @return timestamp of the parsed record
+     */
+    long processRecord(Cursor cursor);
+}
