+++ radar-android-phone-usage/build.gradle
@@ -1,0 +1,14 @@
+apply from: '../gradle/android.gradle'
+
+//---------------------------------------------------------------------------//
+// Configuration                                                             //
+//---------------------------------------------------------------------------//
+
+ext.moduleName = 'radar-android-phone-usage'
+ext.moduleDescription = 'Phone usage statistics plugin for RADAR passive remote monitoring app'
+
+//---------------------------------------------------------------------------//
+// Sources and classpath configurations                                      //
+//---------------------------------------------------------------------------//
+
+apply from: '../gradle/publishing.gradle'
