+++ radar-android-phone-usage/src/main/res/values/strings.xml
@@ -1,0 +1,7 @@
+<resources>
+    <string name="phoneUsageServiceDisplayName">App Usage</string>
+    <string name="app_usage_service_name">%1$s app usage</string>
+    <string name="phone_usage_description">Phone usage measures what apps are used and how long the
+        phone is turned on.
+    </string>
+</resources>
