+++ radar-android-phone/src/main/java/org/radarcns/phone/PhoneState.java
