+++ radar-android-phone/src/main/res/values/strings.xml
@@ -1,0 +1,19 @@
+<resources>
+    <string name="phoneServiceDisplayName">Phone</string>
+    <string name="phoneLocationServiceDisplayName">Location</string>
+    <string name="location_manager_name">%1$s location</string>
+    <string name="phone_location_description">Location is used for estimating distance travelled.
+        Only relative locations are gathered, the actual location is not stored.
+    </string>
+    <string name="phone_contact_list_description">Counts the number of contacts stored on your
+        phone. No information about the contacts themselves are gathered.
+    </string>
+    <string name="phone_bluetooth_description">Assess the number of Bluetooth devices that are
+        within range or connected. No connections are made. Data is only collected if Bluetooth is
+        turned on.
+    </string>
+    <string name="phone_sensors_description">Phone sensors like acceleration and step counts.
+    </string>
+    <string name="contact_list">Contacts List</string>
+    <string name="bluetooth_devices">Bluetooth Devices</string>
+</resources>
