+++ settings.gradle
@@ -1,2 +1,5 @@
-rootProject.name = 'radar-android-phone'
+include ':radar-android-phone'
+include ':radar-android-phone-telephony'
+include ':radar-android-phone-usage'
+
 enableFeaturePreview('STABLE_PUBLISHING')
