--- src/main/AndroidManifest.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.radarcns.phone" >
-
-    <uses-feature android:name="android.hardware.bluetooth" android:required="false"/>
-    <uses-feature android:name="android.hardware.telephony" android:required="false"/>
-    <uses-feature android:name="android.hardware.location" android:required="false"/>
-
-    <uses-permission android:name="android.permission.BLUETOOTH"/>
-    <uses-permission android:name="android.permission.BLUETOOTH_ADMIN"/>
-    <uses-permission android:name="android.permission.READ_CALL_LOG"/>
-    <uses-permission android:name="android.permission.READ_SMS"/>
-    <uses-permission android:name="android.permission.READ_CONTACTS"/>
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
-    <uses-permission android:name="android.permission.PACKAGE_USAGE_STATS"/>
-
-    <application android:allowBackup="true">
-        <service android:name=".PhoneSensorService"/>
-        <service android:name=".PhoneLocationService"/>
-        <service android:name=".PhoneLogService"/>
-        <service android:name=".PhoneUsageService"/>
-        <service android:name=".PhoneContactsListService"/>
-        <service android:name=".PhoneBluetoothService"/>
-    </application>
-</manifest>
